Calculate typeNameToField maps used in EncapsulatedObject::resolve upfront in TypeBuilder
